Moved SummaryMenus handling to InteractableMenu
