+++ RogueEssence/Menu/ScriptableMenu.cs
@@ -14,7 +14,6 @@ namespace RogueEssence.Menu
     public class ScriptableMenu : InteractableMenu
     {
         public List<IMenuElement> MenuElements;
-        public List<SummaryMenu> SummaryMenus;
 
         protected Action<InputManager> UpdateFunction;
 
@@ -25,7 +24,6 @@ namespace RogueEssence.Menu
             UpdateFunction = updateFunction;
             Bounds = new Rect(x, y, w, h);
             MenuElements = new List<IMenuElement>();
-            SummaryMenus = new List<SummaryMenu>();
         }
 
         public override IEnumerable<IMenuElement> GetElements()
@@ -39,23 +37,11 @@ namespace RogueEssence.Menu
             if (UpdateFunction != null)
             UpdateFunction(input);
         }
-
-        public override void Draw(SpriteBatch spriteBatch)
-        {
-            if (!Visible)
-                return;
-            base.Draw(spriteBatch);
-
-            foreach (SummaryMenu menu in SummaryMenus)
-                menu.Draw(spriteBatch);
-        }
     }
 
 
     public class ScriptableSingleStripMenu : SingleStripMenu
     {
-        public List<SummaryMenu> SummaryMenus;
-
         public Action<InputManager> UpdateFunction;
 
         public Action CancelFunction;
@@ -124,8 +110,6 @@ namespace RogueEssence.Menu
 
             int choice_width = CalculateChoiceLength(choices, minWidth);
             Initialize(new Loc(x, y), choice_width, choices.ToArray(), mappedDefault.Value, choices.Count, multiSelect);
-
-            SummaryMenus = new List<SummaryMenu>();
         }
 
         protected override void ChoiceChanged()
@@ -138,17 +122,6 @@ namespace RogueEssence.Menu
         {
             if (MultiSelectChangedFunction != null)
                 MultiSelectChangedFunction();
-        }
-
-
-        public override void Draw(SpriteBatch spriteBatch)
-        {
-            if (!Visible)
-                return;
-            base.Draw(spriteBatch);
-
-            foreach (SummaryMenu menu in SummaryMenus)
-                menu.Draw(spriteBatch);
         }
 
         public override void Update(InputManager input)
@@ -186,8 +159,6 @@ namespace RogueEssence.Menu
 
     public class ScriptableMultiPageMenu : MultiPageMenu
     {
-        public List<SummaryMenu> SummaryMenus;
-
         public Action<InputManager> UpdateFunction;
 
         public Action CancelFunction;
@@ -219,8 +190,6 @@ namespace RogueEssence.Menu
             int defaultPage = defaultTotalChoice / spacesPerPage;
             int defaultChoice = defaultTotalChoice % spacesPerPage;
             Initialize(start, width, title, pagedChoices, defaultChoice, defaultPage, spacesPerPage, showPagesOnSingle, multiSelect);
-
-            SummaryMenus = new List<SummaryMenu>();
         }
 
         protected override void ChoiceChanged()
@@ -233,16 +202,6 @@ namespace RogueEssence.Menu
         {
             if (MultiSelectChangedFunction != null)
                 MultiSelectChangedFunction();
-        }
-
-        public override void Draw(SpriteBatch spriteBatch)
-        {
-            if (!Visible)
-                return;
-            base.Draw(spriteBatch);
-
-            foreach (SummaryMenu menu in SummaryMenus)
-                menu.Draw(spriteBatch);
         }
 
         public override void Update(InputManager input)
